refactor, better organization
