+++ fastplotlib/layouts/_engine.py
@@ -30,14 +30,39 @@ class BaseLayout:
         renderer: pygfx.WgpuRenderer,
         subplots: np.ndarray[Subplot],
         canvas_rect: tuple,
+        moveable: bool,
+        resizeable: bool,
     ):
         self._renderer = renderer
         self._subplots: np.ndarray[Subplot] = subplots.ravel()
         self._canvas_rect = canvas_rect
 
+        self._last_pointer_pos: np.ndarray[np.float64, np.float64] = np.array(
+            [np.nan, np.nan]
+        )
+
+        self._active_action: str | None = None
+        self._active_subplot: Subplot | None = None
+        self._subplot_focus: Subplot | None = None
+
+        for subplot in self._subplots:
+            # highlight plane when pointer enters it
+            subplot.frame.plane.add_event_handler(
+                partial(self._highlight_plane, subplot), "pointer_enter"
+            )
+
+            if resizeable:
+                # highlight/unhighlight resize handler when pointer enters/leaves
+                subplot.frame.resize_handle.add_event_handler(
+                    partial(self._highlight_resize_handler, subplot), "pointer_enter"
+                )
+                subplot.frame.resize_handle.add_event_handler(
+                    partial(self._unhighlight_resize_handler, subplot), "pointer_leave"
+                )
+
     def _inside_render_rect(self, subplot: Subplot, pos: tuple[int, int]) -> bool:
         """whether the pos is within the render area, used for filtering out pointer events"""
-        rect = subplot._fpl_get_render_rect()
+        rect = subplot.frame.get_render_rect()
 
         x0, y0 = rect[:2]
 
@@ -61,10 +86,33 @@ class BaseLayout:
     def set_extent(self, subplot, extent: np.ndarray | list | tuple):
         raise NotImplementedError
 
-    def _fpl_canvas_resized(self, canvas_rect: tuple):
+    def canvas_resized(self, canvas_rect: tuple):
         self._canvas_rect = canvas_rect
         for subplot in self._subplots:
-            subplot._fpl_canvas_resized(canvas_rect)
+            subplot.frame.canvas_resized(canvas_rect)
+
+    def _highlight_resize_handler(self, subplot: Subplot, ev):
+        if self._active_action == "resize":
+            return
+
+        ev.target.material.color = subplot.frame.resize_handle_color.highlight
+
+    def _unhighlight_resize_handler(self, subplot: Subplot, ev):
+        if self._active_action == "resize":
+            return
+
+        ev.target.material.color = subplot.frame.resize_handle_color.idle
+
+    def _highlight_plane(self, subplot: Subplot, ev):
+        if self._active_action is not None:
+            return
+
+        # reset color of previous focus
+        if self._subplot_focus is not None:
+            self._subplot_focus.frame.plane.material.color = subplot.frame.plane_color.idle
+
+        self._subplot_focus = subplot
+        ev.target.material.color = subplot.frame.plane_color.highlight
 
     def __len__(self):
         return len(self._subplots)
@@ -79,7 +127,7 @@ class FlexLayout(BaseLayout):
         moveable=True,
         resizeable=True,
     ):
-        super().__init__(renderer, subplots, canvas_rect)
+        super().__init__(renderer, subplots, canvas_rect, moveable, resizeable)
 
         self._last_pointer_pos: np.ndarray[np.float64, np.float64] = np.array(
             [np.nan, np.nan]
@@ -90,28 +138,14 @@ class FlexLayout(BaseLayout):
         self._subplot_focus: Subplot | None = None
 
         for subplot in self._subplots:
-            # highlight plane when pointer enters it
-            subplot._fpl_plane.add_event_handler(
-                partial(self._highlight_plane, subplot), "pointer_enter"
-            )
-
             if moveable:
                 # start a move action
-                subplot._fpl_plane.add_event_handler(
+                subplot.frame.plane.add_event_handler(
                     partial(self._action_start, subplot, "move"), "pointer_down"
                 )
                 # start a resize action
-                subplot._fpl_resize_handle.add_event_handler(
+                subplot.frame.resize_handle.add_event_handler(
                     partial(self._action_start, subplot, "resize"), "pointer_down"
-                )
-
-            if resizeable:
-                # highlight/unhighlight resize handler when pointer enters/leaves
-                subplot._fpl_resize_handle.add_event_handler(
-                    partial(self._highlight_resize_handler, subplot), "pointer_enter"
-                )
-                subplot._fpl_resize_handle.add_event_handler(
-                    partial(self._unhighlight_resize_handler, subplot), "pointer_leave"
                 )
 
         if moveable or resizeable:
@@ -131,12 +165,12 @@ class FlexLayout(BaseLayout):
         delta_x, delta_y = delta
         if self._active_action == "resize":
             # subtract only from x1, y1
-            new_extent = self._active_subplot.extent - np.asarray(
+            new_extent = self._active_subplot.frame.extent - np.asarray(
                 [0, delta_x, 0, delta_y]
             )
         else:
             # moving
-            new_extent = self._active_subplot.extent - np.asarray(
+            new_extent = self._active_subplot.frame.extent - np.asarray(
                 [delta_x, delta_x, delta_y, delta_y]
             )
 
@@ -147,45 +181,45 @@ class FlexLayout(BaseLayout):
         # make sure width and height are valid
         # min width, height is 50px
         if w <= 50:  # width > 0
-            new_extent[:2] = self._active_subplot.extent[:2]
+            new_extent[:2] = self._active_subplot.frame.extent[:2]
 
         if h <= 50:  # height > 0
-            new_extent[2:] = self._active_subplot.extent[2:]
+            new_extent[2:] = self._active_subplot.frame.extent[2:]
 
         # ignore movement if this would cause an overlap
-        for frame in self._subplots:
-            if frame is self._active_subplot:
+        for subplot in self._subplots:
+            if subplot is self._active_subplot:
                 continue
 
-            if frame.overlaps(new_extent):
+            if subplot.frame.rect_manager.overlaps(new_extent):
                 # we have an overlap, need to ignore one or more deltas
                 # ignore x
-                if not frame.is_left_of(x0) or not frame.is_right_of(x1):
-                    new_extent[:2] = self._active_subplot.extent[:2]
+                if not subplot.frame.rect_manager.is_left_of(x0) or not subplot.frame.rect_manager.is_right_of(x1):
+                    new_extent[:2] = self._active_subplot.frame.extent[:2]
 
                 # ignore y
-                if not frame.is_above(y0) or not frame.is_below(y1):
-                    new_extent[2:] = self._active_subplot.extent[2:]
+                if not subplot.frame.rect_manager.is_above(y0) or not subplot.frame.rect_manager.is_below(y1):
+                    new_extent[2:] = self._active_subplot.frame.extent[2:]
 
         # make sure all vals are non-negative
         if (new_extent[:2] < 0).any():
             # ignore delta_x
-            new_extent[:2] = self._active_subplot.extent[:2]
+            new_extent[:2] = self._active_subplot.frame.extent[:2]
 
         if (new_extent[2:] < 0).any():
             # ignore delta_y
-            new_extent[2:] = self._active_subplot.extent[2:]
+            new_extent[2:] = self._active_subplot.frame.extent[2:]
 
         # canvas extent
         cx0, cy0, cw, ch = self._canvas_rect
 
         # check if new x-range is beyond canvas x-max
         if (new_extent[:2] > cx0 + cw).any():
-            new_extent[:2] = self._active_subplot.extent[:2]
+            new_extent[:2] = self._active_subplot.frame.extent[:2]
 
         # check if new y-range is beyond canvas y-max
         if (new_extent[2:] > cy0 + ch).any():
-            new_extent[2:] = self._active_subplot.extent[2:]
+            new_extent[2:] = self._active_subplot.frame.extent[2:]
 
         return new_extent
 
@@ -196,11 +230,11 @@ class FlexLayout(BaseLayout):
         if ev.button == 1:
             self._active_action = action
             if action == "resize":
-                subplot._fpl_resize_handle.material.color = (
-                    subplot.resize_handle_color.action
+                subplot.frame.resize_handle.material.color = (
+                    subplot.frame.resize_handle_color.action
                 )
             elif action == "move":
-                subplot._fpl_plane.material.color = subplot.plane_color.action
+                subplot.frame.plane.material.color = subplot.frame.plane_color.action
             else:
                 raise ValueError
 
@@ -213,44 +247,21 @@ class FlexLayout(BaseLayout):
 
         delta_x, delta_y = self._last_pointer_pos - (ev.x, ev.y)
         new_extent = self._new_extent_from_delta((delta_x, delta_y))
-        self._active_subplot.extent = new_extent
+        self._active_subplot.frame.extent = new_extent
         self._last_pointer_pos[:] = ev.x, ev.y
 
     def _action_end(self, ev):
         self._active_action = None
         if self._active_subplot is not None:
-            self._active_subplot._fpl_resize_handle.material.color = (
-                self._active_subplot.resize_handle_color.idle
+            self._active_subplot.frame.resize_handle.material.color = (
+                self._active_subplot.frame.resize_handle_color.idle
             )
-            self._active_subplot._fpl_plane.material.color = (
-                self._active_subplot.plane_color.idle
+            self._active_subplot.frame.plane.material.color = (
+                self._active_subplot.frame.plane_color.idle
             )
         self._active_subplot = None
 
         self._last_pointer_pos[:] = np.nan
-
-    def _highlight_resize_handler(self, subplot: Subplot, ev):
-        if self._active_action == "resize":
-            return
-
-        ev.target.material.color = subplot.resize_handle_color.highlight
-
-    def _unhighlight_resize_handler(self, subplot: Subplot, ev):
-        if self._active_action == "resize":
-            return
-
-        ev.target.material.color = subplot.resize_handle_color.idle
-
-    def _highlight_plane(self, subplot: Subplot, ev):
-        if self._active_action is not None:
-            return
-
-        # reset color of previous focus
-        if self._subplot_focus is not None:
-            self._subplot_focus._fpl_plane.material.color = subplot.plane_color.idle
-
-        self._subplot_focus = subplot
-        ev.target.material.color = subplot.plane_color.highlight
 
 
 class GridLayout(FlexLayout):
